Renamed SimpleOptParse to SimpleScripting::Argv
